[BGBUILD-222] Add snapshotting and overwrite support for EBS plugin. Snapshots can be saved from overwrite by setting the 'preserve_snapshots' option to true.
[BGBUILD-193] Vastly improved handling of AWS timing/concurrency issues by using wait-test polling loops to ensure AWS components are in the expected state when actions are performed on them that require a given state.

Gracefully handle scenarios where the state of a component changes between sampling the state and performing an action that becomes invalid. This can be due to an external entity changing something while we're using it.

Gracefully handle orphaned components when using overwrite, this can be because the user deleted some components manually or the build failing previously.